Shane.md
I wanted to say that I wished I could add my phone number, but my phone is out of sorts as of lately since it got factory reset. Also, I am interested in learning not just the programming languages I made for their sake alone, but how they could be applied in making some A.I. of which help robots as well as chatbots operate on websites, and perhaps even the algorithms of the website.